+++ flecsi/exec/leg/tracer.hh
@@ -51,11 +51,6 @@ private:
     }
   }
 
-public:
-  static bool is_tracing() {
-    return tracing;
-  }
-
 private:
   id_t id_;
   bool skip_;